requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
passlib==1.7.4
requests==2.31.0
python-dotenv==1.0.1

google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

